pyproject.toml
[tool.poetry]
name = "rag-notebook"
version = "0.1.0"
description = "Production-style RAG Notebook portfolio project"
authors = ["Sofia Vistas"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
python-multipart = "^0.0.9"
jinja2 = "^3.1.4"
openai = "^1.40.0"
chromadb = "^0.5.5"
pydantic-settings = "^2.4.0"
structlog = "^24.4.0"
pypdf = "^6.7.0"
sqlalchemy = "^2.0.46"
alembic = "^1.18.4"
psycopg = {extras = ["binary"], version = "^3.3.2"}
pgvector = "^0.4.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.11.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.23.8"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
